backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
piexif==1.1.3
motor==3.3.2
pymongo==4.6.0
pydantic[email]==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
httpx==0.25.2
opencv-python-headless==4.10.0.84
numpy==1.26.4
openai==1.52.2
